separated methods for testing
