+++ src/services/elastic.js
@@ -11,7 +11,6 @@ import {
 } from './api/variant/sqon'
 import { elasticSearchTranslator, FILTER_SUBTYPE_NESTED } from './api/variant/sqon/dialect/es'
 
-
 const replacePlaceholderInJSON = ( query, placeholder, placeholderValue ) => {
     return JSON.parse(
         JSON.stringify( query ).split( placeholder ).join( placeholderValue )
@@ -57,6 +56,214 @@ const generateAclFilters = ( acl, service, schema = null ) => {
 
     return filters
 }
+
+export const generateVariantQuery = ( patient, request, acl, schema, group, index, limit ) => {
+    const sortDefinition = schema.groups[ ( !group ? schema.defaultGroup : group ) ]
+    const filter = generateAclFilters( acl, SERVICE_TYPE_VARIANT, schema )
+    let sort = sortDefinition.sort
+
+    if ( sortDefinition.postprocess ) {
+        // eslint-disable-next-line no-new-func
+        const postprocess = new Function( 'context', sortDefinition.postprocess )
+        const context = {
+            sort,
+            acl,
+            patient,
+            index,
+            limit
+        }
+
+        sort = postprocess( context )
+    }
+    if ( filter.length > 0 ) {
+        request.query.bool.filter.push( filter )
+    }
+
+    request.query.bool.filter.push( { term: { [ schema.fields.patient ]: patient } } )
+
+    return {
+        from: index,
+        size: limit,
+        query: replacePlaceholderInJSON( request.query, '%patientId.keyword%', patient ),
+        sort
+    }
+}
+
+export const generateFacetQuery = ( patient, request, denormalizedRequest, acl, schema ) => {
+    const schemaFacets = flatten(
+        map( schema.categories, 'filters' )
+    ).filter( ( filter ) => {
+        return isArray( filter.facet )
+    } )
+    const filter = generateAclFilters( acl, SERVICE_TYPE_VARIANT, schema )
+    const getSearchFieldNameFromFieldId = getFieldSearchNameFromFieldIdMappingFunction( schema )
+    const getFacetFieldNameFromFieldId = getFieldFacetNameFromFieldIdMappingFunction( schema )
+    const getFieldSubtypeFromFieldId = getFieldSubtypeFromFieldIdMappingFunction( schema )
+    const aggs = {
+        filtered: {
+            aggs: {},
+            filter: { bool: { } }
+        }
+    }
+
+    const facetsWithSubtypes = {
+        [ FILTER_SUBTYPE_NESTED ]: {}
+    }
+    const facetFilteredExcept = {}
+
+    // generation of the regular aggregation (all non-nested and all unfiltered one excluded)
+    aggs.filtered.aggs = schemaFacets.reduce( ( accumulator, agg ) => {
+        agg.facet.forEach( ( facet ) => {
+            const facetSubtype = getFieldSubtypeFromFieldId( facet.id )
+
+            if ( !facetSubtype ) {
+                accumulator[ [ facet.id ] ] = facet.query
+            } else {
+                facetsWithSubtypes[ FILTER_SUBTYPE_NESTED ][ [ facet.id ] ] = facetSubtype
+            }
+        } )
+        return accumulator
+    }, {} )
+
+    // generation of the 'filtered' nested aggregations
+
+    if ( facetsWithSubtypes[ FILTER_SUBTYPE_NESTED ] !== {} ) {
+        Object.keys( facetsWithSubtypes[ FILTER_SUBTYPE_NESTED ] ).forEach( ( nestedFacetId ) => {
+            const nestedFacetConfig = facetsWithSubtypes[ FILTER_SUBTYPE_NESTED ][ nestedFacetId ].config
+            const nestedFacetFields = getFacetFieldNameFromFieldId( nestedFacetId )
+
+            if ( !aggs[ [ `nested_${nestedFacetConfig.path}` ] ] ) {
+                aggs[ [ `nested_${nestedFacetConfig.path}` ] ] = {
+                    nested: { path: nestedFacetConfig.path },
+                    aggs: {
+                        filtered: {
+                            filter: {
+                                bool: {
+                                    filter: { term: { [ schema.fields.patient ]: patient } }
+                                }
+                            },
+                            aggs: {}
+                        }
+                    }
+                }
+            }
+            // at the same time generate all nested facet
+            nestedFacetFields.forEach( ( nestedFacetField ) => {
+                aggs[ [ `nested_${nestedFacetConfig.path}` ] ].aggs.filtered.aggs[ nestedFacetField.id ] = nestedFacetField.query
+                if ( !facetFilteredExcept[ [ nestedFacetField.id ] ] ) {
+                    facetFilteredExcept[ [ nestedFacetField.id ] ] = { [ `nested_${nestedFacetConfig.path}` ]: aggs[ [ `nested_${nestedFacetConfig.path}` ] ] }
+                }
+            } )
+        } )
+    }
+    // generate unfiltered aggs
+    traverseArrayAndApplyFunc( denormalizedRequest.instructions, ( index, instruction ) => {
+        if ( instructionIsFilter( instruction ) ) {
+            const facetId = instruction.data.id
+            const facetFields = getFacetFieldNameFromFieldId( facetId )
+
+            if ( facetFields ) {
+                let searchFields = getSearchFieldNameFromFieldId( facetId )
+
+                // boolean comparison -- normalize to always be an object
+                searchFields = isString( searchFields ) ? { [ facetId ]: searchFields } : searchFields
+
+                Object.keys( searchFields ).forEach( ( facetField ) => {
+                    let filteredExceptAggs = { [ facetField ]: aggs.filtered.aggs[ facetField ] }
+                    let instructionsWithoutFacetId = []
+
+                    traverseArrayAndApplyFunc( denormalizedRequest.instructions, ( iindex, iinstruction ) => {
+                        if ( !isArray( iinstruction ) ) {
+                            if ( !instructionIsFilter( iinstruction ) ) {
+                                instructionsWithoutFacetId.push( iinstruction )
+                            } else {
+                                const fiinstruction = cloneDeep( iinstruction )
+
+                                if ( iinstruction.data.id === facetId ) {
+                                    fiinstruction.data.values = []
+                                }
+                                instructionsWithoutFacetId.push( fiinstruction )
+                            }
+                        }
+                    } )
+
+                    const translatedFacet = elasticSearchTranslator.translate( { instructions: instructionsWithoutFacetId }, {}, getSearchFieldNameFromFieldId, getFacetFieldNameFromFieldId, getFieldSubtypeFromFieldId )
+                    const normalizedNestedFields = Object.keys( facetsWithSubtypes[ FILTER_SUBTYPE_NESTED ] ).filter( ( key ) => key.replace( '_min', '' ).replace( '_max', '' ) === facetField )
+                    const isNestedAgg = normalizedNestedFields.length > 0
+
+                    if ( isNestedAgg ) {
+
+                        normalizedNestedFields.forEach( ( normalizedNestedField ) => {
+                            filteredExceptAggs = cloneDeep( facetFilteredExcept[ normalizedNestedField ] )
+                            const normalizedNestedPath = Object.keys( filteredExceptAggs )[ 0 ]
+
+                            filteredExceptAggs[ normalizedNestedPath ].aggs.filtered.aggs = facetFields.reduce( ( fieldAcc, facetFieldData ) => {
+                                fieldAcc[ [ facetFieldData.id ] ] = facetFieldData.query
+                                return fieldAcc
+                            }, {} )
+                        } )
+                    } else {
+
+                        // non nested aggs min max should be here
+                        if ( JSON.stringify( filteredExceptAggs ) === '{}' ) {
+                            // min and max
+                            filteredExceptAggs = { [ `${facetField}_min` ]: aggs.filtered.aggs[ `${facetField}_min` ],
+                                [ `${facetField}_max` ]: aggs.filtered.aggs[ `${facetField}_max` ]
+                            }
+                        }
+
+                        translatedFacet.query.bool.filter.push( { bool: {
+                            filter: { term: { [ schema.fields.patient ]: patient } }
+                        } } )
+                    }
+
+                    aggs[ [ facetField ] ] = {
+                        global: {},
+                        aggs: {
+                            [ `filtered_except_${facetField}` ]: {
+                                filter: translatedFacet.query,
+                                aggs: filteredExceptAggs
+                            }
+                        }
+                    }
+                } )
+            }
+        }
+    } )
+
+    const query = {
+        bool: {
+            filter: request.query.bool.filter
+        }
+    }
+
+    if ( filter.length > 0 ) {
+        query.bool.filter.push( filter )
+    }
+
+    query.bool.filter.push( { term: { [ schema.fields.patient ]: patient } } )
+
+    return {
+        size: 0,
+        query: replacePlaceholderInJSON( query, '%patientId.keyword%', patient ),
+        aggs: replacePlaceholderInJSON( aggs, '%patientId.keyword%', patient )
+    }
+}
+
+export const generateCountQuery = ( patient, request, acl, schema ) => {
+    const filter = generateAclFilters( acl, SERVICE_TYPE_VARIANT, schema )
+
+    if ( filter.length > 0 ) {
+        request.query.bool.filter.push( filter )
+    }
+
+    request.query.bool.filter.push( { term: { [ schema.fields.patient ]: patient } } )
+
+    return {
+        query: replacePlaceholderInJSON( request.query, '%patientId.keyword%', patient )
+    }
+}
+
 
 export default class ElasticClient {
 
@@ -141,35 +348,7 @@ export default class ElasticClient {
 
     async searchVariantsForPatient( patient, request, acl, schema, group, index, limit ) {
         const uri = `${this.host}${schema.path}/_search`
-        const sortDefinition = schema.groups[ ( !group ? schema.defaultGroup : group ) ]
-        const filter = generateAclFilters( acl, SERVICE_TYPE_VARIANT, schema )
-        let sort = sortDefinition.sort
-
-        if ( sortDefinition.postprocess ) {
-            // eslint-disable-next-line no-new-func
-            const postprocess = new Function( 'context', sortDefinition.postprocess )
-            const context = {
-                sort,
-                acl,
-                patient,
-                index,
-                limit
-            }
-
-            sort = postprocess( context )
-        }
-        if ( filter.length > 0 ) {
-            request.query.bool.filter.push( filter )
-        }
-
-        request.query.bool.filter.push( { term: { [ schema.fields.patient ]: patient } } )
-
-        const body = {
-            from: index,
-            size: limit,
-            query: replacePlaceholderInJSON( request.query, '%patientId.keyword%', patient ),
-            sort
-        }
+        const body = generateVariantQuery( patient, request, acl, schema, group, index, limit )
 
         console.debug( `searchVariantsForPatient: ${JSON.stringify( body )}` )
 
@@ -183,164 +362,7 @@ export default class ElasticClient {
 
     async getFacetsForVariant( patient, request, denormalizedRequest, acl, schema ) {
         const uri = `${this.host}${schema.path}/_search`
-        const schemaFacets = flatten(
-            map( schema.categories, 'filters' )
-        ).filter( ( filter ) => {
-            return isArray( filter.facet )
-        } )
-        const filter = generateAclFilters( acl, SERVICE_TYPE_VARIANT, schema )
-        const getSearchFieldNameFromFieldId = getFieldSearchNameFromFieldIdMappingFunction( schema )
-        const getFacetFieldNameFromFieldId = getFieldFacetNameFromFieldIdMappingFunction( schema )
-        const getFieldSubtypeFromFieldId = getFieldSubtypeFromFieldIdMappingFunction( schema )
-        const aggs = {
-            filtered: {
-                aggs: {},
-                filter: { bool: { } }
-            }
-        }
-
-        const facetsWithSubtypes = {
-            [ FILTER_SUBTYPE_NESTED ]: {}
-        }
-        const facetFilteredExcept = {}
-
-        // generation of the regular aggregation (all non-nested and all unfiltered one excluded)
-        aggs.filtered.aggs = schemaFacets.reduce( ( accumulator, agg ) => {
-            agg.facet.forEach( ( facet ) => {
-                const facetSubtype = getFieldSubtypeFromFieldId( facet.id )
-
-                if ( !facetSubtype ) {
-                    accumulator[ [ facet.id ] ] = facet.query
-                } else {
-                    facetsWithSubtypes[ FILTER_SUBTYPE_NESTED ][ [ facet.id ] ] = facetSubtype
-                }
-            } )
-            return accumulator
-        }, {} )
-
-        // generation of the 'filtered' nested aggregations
-
-        if ( facetsWithSubtypes[ FILTER_SUBTYPE_NESTED ] !== {} ) {
-            Object.keys( facetsWithSubtypes[ FILTER_SUBTYPE_NESTED ] ).forEach( ( nestedFacetId ) => {
-                const nestedFacetConfig = facetsWithSubtypes[ FILTER_SUBTYPE_NESTED ][ nestedFacetId ].config
-                const nestedFacetFields = getFacetFieldNameFromFieldId( nestedFacetId )
-
-                if ( !aggs[ [ `nested_${nestedFacetConfig.path}` ] ] ) {
-                    aggs[ [ `nested_${nestedFacetConfig.path}` ] ] = {
-                        nested: { path: nestedFacetConfig.path },
-                        aggs: {
-                            filtered: {
-                                filter: {
-                                    bool: {
-                                        filter: { term: { [ schema.fields.patient ]: patient } }
-                                    }
-                                },
-                                aggs: {}
-                            }
-                        }
-                    }
-                }
-                // at the same time generate all nested facet
-                nestedFacetFields.forEach( ( nestedFacetField ) => {
-                    aggs[ [ `nested_${nestedFacetConfig.path}` ] ].aggs.filtered.aggs[ nestedFacetField.id ] = nestedFacetField.query
-                    if ( !facetFilteredExcept[ [ nestedFacetField.id ] ] ) {
-                        facetFilteredExcept[ [ nestedFacetField.id ] ] = { [ `nested_${nestedFacetConfig.path}` ]: aggs[ [ `nested_${nestedFacetConfig.path}` ] ] }
-                    }
-                } )
-            } )
-        }
-        // generate unfiltered aggs
-        traverseArrayAndApplyFunc( denormalizedRequest.instructions, ( index, instruction ) => {
-            if ( instructionIsFilter( instruction ) ) {
-                const facetId = instruction.data.id
-                const facetFields = getFacetFieldNameFromFieldId( facetId )
-
-                if ( facetFields ) {
-                    let searchFields = getSearchFieldNameFromFieldId( facetId )
-
-                    // boolean comparison -- normalize to always be an object
-                    searchFields = isString( searchFields ) ? { [ facetId ]: searchFields } : searchFields
-
-                    Object.keys( searchFields ).forEach( ( facetField ) => {
-                        let filteredExceptAggs = { [ facetField ]: aggs.filtered.aggs[ facetField ] }
-                        let instructionsWithoutFacetId = []
-
-                        traverseArrayAndApplyFunc( denormalizedRequest.instructions, ( iindex, iinstruction ) => {
-                            if ( !isArray( iinstruction ) ) {
-                                if ( !instructionIsFilter( iinstruction ) ) {
-                                    instructionsWithoutFacetId.push( iinstruction )
-                                } else {
-                                    const fiinstruction = cloneDeep( iinstruction )
-
-                                    if ( iinstruction.data.id === facetId ) {
-                                        fiinstruction.data.values = []
-                                    }
-                                    instructionsWithoutFacetId.push( fiinstruction )
-                                }
-                            }
-                        } )
-
-                        const translatedFacet = elasticSearchTranslator.translate( { instructions: instructionsWithoutFacetId }, {}, getSearchFieldNameFromFieldId, getFacetFieldNameFromFieldId, getFieldSubtypeFromFieldId )
-                        const normalizedNestedFields = Object.keys( facetsWithSubtypes[ FILTER_SUBTYPE_NESTED ] ).filter( ( key ) => key.replace( '_min', '' ).replace( '_max', '' ) === facetField )
-                        const isNestedAgg = normalizedNestedFields.length > 0
-
-                        if ( isNestedAgg ) {
-
-                            normalizedNestedFields.forEach( ( normalizedNestedField ) => {
-                                filteredExceptAggs = cloneDeep( facetFilteredExcept[ normalizedNestedField ] )
-                                const normalizedNestedPath = Object.keys( filteredExceptAggs )[ 0 ]
-
-                                filteredExceptAggs[ normalizedNestedPath ].aggs.filtered.aggs = facetFields.reduce( ( fieldAcc, facetFieldData ) => {
-                                    fieldAcc[ [ facetFieldData.id ] ] = facetFieldData.query
-                                    return fieldAcc
-                                }, {} )
-                            } )
-                        } else {
-
-                            // non nested aggs min max should be here
-                            if ( JSON.stringify( filteredExceptAggs ) === '{}' ) {
-                                // min and max
-                                filteredExceptAggs = { [ `${facetField}_min` ]: aggs.filtered.aggs[ `${facetField}_min` ],
-                                    [ `${facetField}_max` ]: aggs.filtered.aggs[ `${facetField}_max` ]
-                                }
-                            }
-
-                            translatedFacet.query.bool.filter.push( { bool: {
-                                filter: { term: { [ schema.fields.patient ]: patient } }
-                            } } )
-                        }
-
-                        aggs[ [ facetField ] ] = {
-                            global: {},
-                            aggs: {
-                                [ `filtered_except_${facetField}` ]: {
-                                    filter: translatedFacet.query,
-                                    aggs: filteredExceptAggs
-                                }
-                            }
-                        }
-                    } )
-                }
-            }
-        } )
-
-        const query = {
-            bool: {
-                filter: request.query.bool.filter
-            }
-        }
-
-        if ( filter.length > 0 ) {
-            query.bool.filter.push( filter )
-        }
-
-        query.bool.filter.push( { term: { [ schema.fields.patient ]: patient } } )
-
-        const body = {
-            size: 0,
-            query: replacePlaceholderInJSON( query, '%patientId.keyword%', patient ),
-            aggs: replacePlaceholderInJSON( aggs, '%patientId.keyword%', patient )
-        }
+        const body = generateFacetQuery( patient, request, denormalizedRequest, acl, schema )
 
         console.debug( `getFacetsForVariant: ${JSON.stringify( body )}` )
 
@@ -354,32 +376,7 @@ export default class ElasticClient {
 
     async countVariantsForPatient( patient, request, acl, schema, group ) {
         const uri = `${this.host}${schema.path}/_count`
-        const filter = generateAclFilters( acl, SERVICE_TYPE_VARIANT, schema )
-        const sortDefinition = schema.groups[ ( !group ? schema.defaultGroup : group ) ]
-        let sort = sortDefinition.sort
-
-        if ( sortDefinition.postprocess ) {
-            const postprocess = new Function( 'context', sortDefinition.postprocess ) /* eslint-disable-line */
-            const context = {
-                sort,
-                acl,
-                patient,
-                index: 0,
-                limit: 0
-            }
-
-            sort = postprocess( context )
-        }
-
-        if ( filter.length > 0 ) {
-            request.query.bool.filter.push( filter )
-        }
-
-        request.query.bool.filter.push( { term: { [ schema.fields.patient ]: patient } } )
-
-        const body = {
-            query: replacePlaceholderInJSON( request.query, '%patientId.keyword%', patient )
-        }
+        const body = generateCountQuery( patient, request, acl, schema, group )
 
         console.debug( `countVariantsForPatient: ${JSON.stringify( body )}` )
 
@@ -387,8 +384,7 @@ export default class ElasticClient {
             method: 'POST',
             uri,
             json: true,
-            body,
-            sort
+            body
         } )
     }
